fix: components/elementsにheaderドメインを含めるように・headerをstyled-components化
