+++ components/elements/header/index.tsx
@@ -1,0 +1,11 @@
+import { GlobalStyle } from "./styles";
+
+const Header: React.FC = () => {
+  return (
+    <>
+      <GlobalStyle />
+      <div className="header"></div>
+    </>
+  );
+};
+export default Header;
